Refactor TLS check (#8792)
* Refactor TLS check, split into two separate classes
