+++ tls/tests/test_remote.py
@@ -6,7 +6,19 @@ import pytest
 from six import PY2
 
 from datadog_checks.base import ConfigurationError
-from datadog_checks.tls import TLSCheck
+from datadog_checks.tls.const import (
+    SERVICE_CHECK_CAN_CONNECT,
+    SERVICE_CHECK_EXPIRATION,
+    SERVICE_CHECK_VALIDATION,
+    SERVICE_CHECK_VERSION,
+)
+from datadog_checks.tls.tls import TLSCheck
+from datadog_checks.tls.tls_remote import TLSRemoteCheck
+
+
+def test_right_class_is_instantiated(instance_remote_no_server):
+    c = TLSCheck('tls', {}, [instance_remote_no_server])
+    assert isinstance(c, TLSRemoteCheck)
 
 
 def test_no_server(instance_remote_no_server):
@@ -20,10 +32,10 @@ def test_ok(aggregator, instance_remote_ok):
     c = TLSCheck('tls', {}, [instance_remote_ok])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -34,10 +46,10 @@ def test_ok_ip(aggregator, instance_remote_ok_ip):
     c = TLSCheck('tls', {}, [instance_remote_ok_ip])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -48,10 +60,10 @@ def test_ok_udp(aggregator, instance_remote_ok_udp):
     c = TLSCheck('tls', {}, [instance_remote_ok_udp])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -62,13 +74,13 @@ def test_no_resolve(aggregator, instance_remote_no_resolve):
     c = TLSCheck('tls', {}, [instance_remote_no_resolve])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, count=0)
 
     message = 'Unable to resolve host, check your DNS'
-    assert message in aggregator.service_checks(c.SERVICE_CHECK_CAN_CONNECT)[0].message
+    assert message in aggregator.service_checks(SERVICE_CHECK_CAN_CONNECT)[0].message
 
     aggregator.assert_all_metrics_covered()
 
@@ -77,13 +89,13 @@ def test_no_connect(aggregator, instance_remote_no_connect):
     c = TLSCheck('tls', {}, [instance_remote_no_connect])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, count=0)
 
     message = 'Unable to resolve host, check your DNS'
-    assert message not in aggregator.service_checks(c.SERVICE_CHECK_CAN_CONNECT)[0].message
+    assert message not in aggregator.service_checks(SERVICE_CHECK_CAN_CONNECT)[0].message
 
     aggregator.assert_all_metrics_covered()
 
@@ -92,13 +104,13 @@ def test_no_connect_port_in_host(aggregator, instance_remote_no_connect_port_in_
     c = TLSCheck('tls', {}, [instance_remote_no_connect_port_in_host])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, count=0)
 
     message = 'Unable to resolve host, check your DNS'
-    assert message not in aggregator.service_checks(c.SERVICE_CHECK_CAN_CONNECT)[0].message
+    assert message not in aggregator.service_checks(SERVICE_CHECK_CAN_CONNECT)[0].message
 
     aggregator.assert_all_metrics_covered()
 
@@ -109,15 +121,15 @@ def test_no_connect_ipv6(aggregator, instance_remote_no_connect):
         c.check(None)
 
     aggregator.assert_service_check(
-        c.SERVICE_CHECK_CAN_CONNECT,
+        SERVICE_CHECK_CAN_CONNECT,
         status=c.CRITICAL,
         tags=c._tags,
         message='No valid addresses found, try checking your IPv6 connectivity',
         count=1,
     )
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, count=0)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, count=0)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, count=0)
 
     aggregator.assert_all_metrics_covered()
 
@@ -126,10 +138,10 @@ def test_version_default_1_1(aggregator, instance_remote_version_default_1_1):
     c = TLSCheck('tls', {}, [instance_remote_version_default_1_1])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.CRITICAL, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -140,10 +152,10 @@ def test_version_default_1_2(aggregator, instance_remote_version_default_1_2):
     c = TLSCheck('tls', {}, [instance_remote_version_default_1_2])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -154,10 +166,10 @@ def test_version_default_1_3(aggregator, instance_remote_version_default_1_3):
     c = TLSCheck('tls', {}, [instance_remote_version_default_1_3])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -168,10 +180,10 @@ def test_version_init_config_default(aggregator, instance_remote_version_default
     c = TLSCheck('tls', {'allowed_versions': ['1.1']}, [instance_remote_version_default_1_1])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -182,10 +194,10 @@ def test_hostname_mismatch(aggregator, instance_remote_hostname_mismatch):
     c = TLSCheck('tls', {}, [instance_remote_hostname_mismatch])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.CRITICAL, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -196,10 +208,10 @@ def test_self_signed_ok(aggregator, instance_remote_self_signed_ok):
     c = TLSCheck('tls', {}, [instance_remote_self_signed_ok])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -210,15 +222,15 @@ def test_cert_expired(aggregator, instance_remote_cert_expired):
     c = TLSCheck('tls', {}, [instance_remote_cert_expired])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.CRITICAL, tags=c._tags, count=1)
     if PY2:
-        aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
-        aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, count=0)
+        aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
+        aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, count=0)
     else:
-        aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, count=0)
+        aggregator.assert_service_check(SERVICE_CHECK_VERSION, count=0)
         aggregator.assert_service_check(
-            c.SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, message='Certificate has expired', count=1
+            SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, message='Certificate has expired', count=1
         )
 
     aggregator.assert_all_metrics_covered()
@@ -228,10 +240,10 @@ def test_fetch_intermediate_certs(aggregator, instance_remote_fetch_intermediate
     c = TLSCheck('tls', {}, [instance_remote_fetch_intermediate_certs])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.OK, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -242,10 +254,10 @@ def test_cert_critical_days(aggregator, instance_remote_cert_critical_days):
     c = TLSCheck('tls', {}, [instance_remote_cert_critical_days])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -256,10 +268,10 @@ def test_cert_critical_seconds(aggregator, instance_remote_cert_critical_seconds
     c = TLSCheck('tls', {}, [instance_remote_cert_critical_seconds])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.CRITICAL, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -270,10 +282,10 @@ def test_cert_warning_days(aggregator, instance_remote_cert_warning_days):
     c = TLSCheck('tls', {}, [instance_remote_cert_warning_days])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.WARNING, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.WARNING, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
@@ -284,10 +296,10 @@ def test_cert_warning_seconds(aggregator, instance_remote_cert_warning_seconds):
     c = TLSCheck('tls', {}, [instance_remote_cert_warning_seconds])
     c.check(None)
 
-    aggregator.assert_service_check(c.SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
-    aggregator.assert_service_check(c.SERVICE_CHECK_EXPIRATION, status=c.WARNING, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_CAN_CONNECT, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VERSION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_VALIDATION, status=c.OK, tags=c._tags, count=1)
+    aggregator.assert_service_check(SERVICE_CHECK_EXPIRATION, status=c.WARNING, tags=c._tags, count=1)
 
     aggregator.assert_metric('tls.days_left', count=1)
     aggregator.assert_metric('tls.seconds_left', count=1)
